requirements.txt
# Azure Storage SDK instead of boto3
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
colorlog==6.8.0
duckdb==1.1.3
fsspec>=2024.2.0
git+https://github.com/ibis-project/ibis.git@bd30ba5#egg=ibis-framework[duckdb]
ipython==8.12.0  # Pinned to avoid compatibility issues with SQLMesh
sqlfluff==3.0.0
sqlglot==26.2.1
sqlmesh[web]==0.146.0
numpy==1.26.4  # Pinning to numpy 1.x for compatibility
pandas<3.0.0  # Ensuring pandas is compatible with numpy 1.x
python-dotenv>=1.0.1
adlfs>=2024.2.0  # Azure Data Lake filesystem for fsspec
ipywidgets>=8.0.0  # Required for IPython/SQLMesh compatibility

# Security and Authentication
PyJWT>=2.8.0
Flask>=2.3.0
Werkzeug>=2.3.0
cryptography>=41.0.0

# Database Connections
psycopg2-binary==2.9.9

# Development and Linting Dependencies
black==24.1.1
isort==5.13.2
ruff>=0.3.0